Teste VP

+++ include/Book.h
@@ -18,7 +18,7 @@ class Book
         string get_isbn13();
         string get_title();
         int get_bestsellers_rank();
-        int get_id();
+        long int get_id();
         int get_rating_count();
         float get_rating_avg();
 
@@ -27,7 +27,7 @@ class Book
         void set_categories(string categories);
         void set_description(string description);
         void set_edition(string edition);
-        void set_id(int id);
+        void set_id(long int id);
         void set_isbn10(string isbn10);
         void set_isbn13(string isbn13);
         void set_rating_avg(float rating_avg);
@@ -40,7 +40,7 @@ class Book
         string categories;
         string description;
         string edition;
-        int id;
+        long int id;
         string isbn10;
         string isbn13;
         float rating_avg;
